fix: resolve circular dependency by moving provider setup types to Meridian.Ui.Shared
IProviderModuleSetupService and ProviderModuleSetupModels were defined in
Meridian.Ui.Services which already references Meridian.Ui.Shared — creating a
circular dependency when Meridian.Ui.Shared tried to import them. Move both
files to Meridian.Ui.Shared.Services and update all consumers.

Also fixes:
- Mutation endpoints now return ProviderModuleSetupResult (success/error) instead
  of {moduleId} anonymous object — matching the TypeScript client expectation
- AlpacaHistoricalDataProvider and AlpacaSymbolSearchProviderRefactored now receive
  credentials from the module context instead of relying solely on env vars
